rename package; tidy up pyproject
