+++ README.md
@@ -8,5 +8,6 @@ To do:
 - [ ] Implement [exponential map flows](https://arxiv.org/abs/2002.02428)
 - [ ] Implement [bump layers](https://arxiv.org/abs/2110.00351)
 - [ ] Extend models, sampling, and some visualisation to $D > 2$
-- [ ] Add documentation and examples
+- [ ] Add documentation and references
+- [ ] Add installation and usage instructions and example outputs in this README
 - [ ] Integrate working models into [torchnf](https://github.com/marshrossney/torchnf) package
