+++ examples/circle_flow/mobius/config.yaml
@@ -1,15 +1,15 @@
 target:
-  class_path: vonmises.distributions.VonMisesFisherMixtureDensity
+  class_path: flows_on_spheres.distributions.VonMisesFisherMixtureDensity
   init_args:
     κ: [10, 10, 10, 10]
     μ: [[1, 1], [1, -1], [-1, -1], [-1, 1]]
     weights: [1, 1, 1, 1]
 
 flow:
-  class_path: vonmises.flows.CircularFlow
+  class_path: flows_on_spheres.flows.CircularFlow
   init_args:
     transformer:
-      class_path: vonmises.transforms.MobiusMixtureTransform
+      class_path: flows_on_spheres.transforms.MobiusMixtureTransform
       init_args:
         n_mixture: 12
     n_layers: 1
