+++ examples/circle_flow/spline/config.yaml
@@ -1,18 +1,18 @@
 target:
-  class_path: vonmises.distributions.VonMisesFisherMixtureDensity
+  class_path: flows_on_spheres.distributions.VonMisesFisherMixtureDensity
   init_args:
     κ: [10, 10, 10, 10]
     μ: [[1, 1], [1, -1], [-1, -1], [-1, 1]]
     weights: [1, 1, 1, 1]
 
 flow: 
-  class_path: vonmises.flows.CircularFlow
+  class_path: flows_on_spheres.flows.CircularFlow
   init_args:
     transformer:
-      class_path: vonmises.transforms.CircularRQSplineTransform
+      class_path: flows_on_spheres.transforms.CircularRQSplineTransform
       init_args:
         n_segments: 12
-    n_layers: 1
+    n_layers: 2
 
 steps: 4000
 batch_size: 4000
