+++ examples/recursive_flow_s2/config.yaml
@@ -1,19 +1,19 @@
 target:
-  class_path: vonmises.distributions.VonMisesFisherMixtureDensity
+  class_path: flows_on_spheres.distributions.VonMisesFisherMixtureDensity
   init_args:
     κ: [10, 10, 10]
     μ: [[0, 1, 0], [0, -1, 0], [0, 0, 1]]
     weights: [1, 1, 1]
 
 flow:
-  class_path: vonmises.flows.RecursiveFlowS2
+  class_path: flows_on_spheres.flows.RecursiveFlowS2
   init_args:
     z_transformer:
-      class_path: vonmises.transforms.BSplineTransform
+      class_path: flows_on_spheres.transforms.BSplineTransform
       init_args:
         n_segments: 12
     xy_transformer:
-      class_path: vonmises.transforms.MobiusMixtureTransform
+      class_path: flows_on_spheres.transforms.MobiusMixtureTransform
       init_args:
         n_mixture: 12
         weighted: true
